+++ src/main/java/spring/springserver/domain/post/data/response/PostMemberReponse.kt
@@ -1,0 +1,26 @@
+package spring.springserver.domain.post.data.response
+
+import spring.springserver.domain.member.entity.Member
+
+data class PostMemberResponse(
+    val id: Long?,
+
+    val username: String,
+
+    val name: String,
+) {
+
+    companion object {
+
+        fun of(
+            member: Member
+        ): PostMemberResponse {
+
+            return PostMemberResponse(
+                member.getId(),
+                member.username,
+                member.name
+            )
+        }
+    }
+}
