+++ src/main/java/spring/springserver/domain/post/data/response/PostResponse.kt
@@ -38,25 +38,3 @@ data class PostResponse(
     }
 }
 
-data class PostMemberResponse(
-    val id: Long?,
-
-    val username: String,
-
-    val name: String,
-) {
-
-    companion object {
-
-        fun of(
-            member: spring.springserver.domain.member.entity.Member
-        ): PostMemberResponse {
-
-            return PostMemberResponse(
-                member.getId(),
-                member.username,
-                member.name
-            )
-        }
-    }
-}
